+++ src/components/ids-data-grid/demos/multiple-virtual-scroll.ts
@@ -12,6 +12,11 @@ const url: any = productsJSON;
 const columns1: IdsDataGridColumn[] = [];
 const dataGrid1 = document.querySelector<IdsDataGrid>('#data-grid-1')!;
 
+const button = document.querySelector('#button-1')!;
+
+const lists: IdsDataGrid[] = [];
+lists.push(dataGrid1);
+
 (async function init() {
   // Set up columns
   columns1.push({
@@ -50,7 +55,14 @@ const dataGrid1 = document.querySelector<IdsDataGrid>('#data-grid-1')!;
     name: 'In Stock',
     field: 'inStock',
     formatter: dataGrid1.formatters.text,
-    sortable: true
+    filterType: dataGrid1.filters.dropdown,
+    filterConditions: [
+      { value: 'not-filtered', label: ' ' },
+      { value: 'True', label: 'True' },
+      { value: 'False', label: 'False' },
+      { value: 'True', label: 'true' },
+      { value: 'False', label: 'false' },
+    ]
   });
   columns1.push({
     id: 'productId',
@@ -176,5 +188,12 @@ const dataGrid2 = document.querySelector<IdsDataGrid>('#data-grid-2')!;
     dataGrid2.data = data;
   };
 
-  setData();
+  await setData();
+  button?.addEventListener('click', async () => {
+    await setData();
+  });
+
+  dataGrid1.addEventListener('selected', (e: any) => {
+    console.info('selected', e.detail);
+  });
 }());